requirements.txt
pandas
numpy
torch
faiss-cpu
tqdm
ollama
langchain-core
transformers
pillow
ipython